Create Phising_Websites.py
This is a project that detects malicious (also known as phising) and non-malicious websites using a public dataset. The algorithms that were used in this project were the Decision Tree algorithm and the k- Nearest Neighbours (k-NN). To train the algorithms i used the train-split method. Also, i performed normalization for k-NN for maximum accuracy and finding of the optimal solution.
You can find the dataset here: https://archive.ics.uci.edu/dataset/327/phishing+websites